Move sound feedback setting to Controllers
Move the sound feedback setting from the Display section to
the Controllers section in the Settings, next to the one for
enabling and disabling haptic feedback. This will make this
functionality more discoverable.

Slightly increase the height of the Controllers section.
